+++ NEWS.md
@@ -1,3 +1,9 @@
+# gglgbtq 0.1.1
+
+* Implemented the following palettes: `agender`, `alloace`, `aroallo`, `bigender`, `demiboy`, `demigirl`.
+* Changed `gay_man` to `gay` (and, analogously, `gay_7`).
+* Moved palettes to the "Gallery" vignette.
+
 # gglgbtq 0.1.0
 
 * Initial version of the package.